ファイル構成の修正
