+++ packages/kokoas-server/src/assets/getFilePath.ts
@@ -8,7 +8,8 @@ type FileName =
 | '工事請負契約約款'
 | '請求書'
 | '原価見積'
-| '工事依頼書';
+| '工事依頼書'
+| '工事依頼書請負書';
 
 
 export const getFilePath = ({
