+++ packages/kokoas-server/src/handleRequest/reqDownloadOrderToSupplier/index.ts
@@ -1,3 +1,3 @@
-export * from './createOrderDocument';
+export * from './createOrderDocment/createOrderDocument';
 export * from './downloadOrderToSupplier';
 export * from './getOrderData';
